requirements.txt
django==3.2.25
flake8==6.1.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
django-split-settings==1.1.0
django-debug-toolbar==3.2.4
